apps/api/pyproject.toml
[tool.poetry]
name = "founder-os-api"
version = "0.1.0"
description = "FounderOS backend API — knowledge base, milestones, accountability diary"
authors = ["Spenser Courville-Taylor"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pydantic = {extras = ["email"], version = "^2.12.5"}
pydantic-settings = "^2.6.0"
pgvector = "^0.3.6"
openai = "^1.57.0"
python-multipart = "^0.0.18"
httpx = "^0.28.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
unstructured = {extras = ["pdf"], version = "^0.16.0"}
tiktoken = "^0.8.0"
sse-starlette = "^2.1.0"
supabase = "^2.11.0"
greenlet = "^3.3.2"
gunicorn = "^25.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
coverage = "^7.6.0"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
concurrency = ["greenlet", "thread"]

[tool.coverage.report]
show_missing = true
fail_under = 90

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

